First implementation of SSAO, kind of works?

+++ Mizu/src/render_core/rhi/rhi_helpers.cpp
@@ -51,7 +51,7 @@ void RHIHelpers::set_material(RenderCommandBuffer& command,
     }
 }
 
-glm::uvec3 RHIHelpers::compute_group_count(glm::uvec3 thread_count, glm::vec3 group_size)
+glm::uvec3 RHIHelpers::compute_group_count(glm::uvec3 thread_count, glm::uvec3 group_size)
 {
     return {(thread_count.x + group_size.x - 1) / group_size.x,
             (thread_count.y + group_size.y - 1) / group_size.y,
